pyproject.toml
[project]
name = "my-newsletters"
version = "0.1.0"
description = "AI-powered voice assistant for daily newsletter briefings"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    # Core async web framework
    "quart>=0.19.0",
    "quart-cors>=0.7.0",
    
    # Database and ORM
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    
    # Data validation and configuration
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    
    # AI and Voice Processing
    "vocode>=0.1.108",
    "elevenlabs>=0.2.26",
    "openai>=1.0.0",
    
    # Gmail and Email Processing
    "google-api-python-client>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.2.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    
    # HTTP and async utilities
    "httpx>=0.25.0",
    "aiofiles>=23.0.0",
    "asyncio-mqtt>=0.16.0",
    
    # JWT and authentication
    "pyjwt>=2.8.0",
    "cryptography>=41.0.0",
    
    # Utility libraries
    "uuid>=1.30",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    
    # Code formatting and linting
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    
    # Type stubs
    "types-python-dateutil",
    "types-requests",
    "types-beautifulsoup4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | myenv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py313"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "unit: marks tests as unit tests",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["backend"]
omit = [
    "*/tests/*",
    "*/venv/*",
    "*/myenv/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]